# Source/OLLAMACHAT.Generated/Controllers/CParserApi.cs
namespace OLLAMACHAT.Generated.Controllers;

/// <summary>
/// </summary>
[ApiController]
[Route("/api/v1/parser/csharp/")]
public class CParserApiController(
    IMediator mediator,
    IMapperInterface mapper,
    IOptions<SolutionSettings> slnSettings,
    ILogger<CParserApiController> logger) : ControllerBase
{
    private string SolutionDirectory => Path.GetDirectoryName(slnSettings.Value.SolutionFilePath)!;

    /// <summary>
    /// Get initialization file names
    /// </summary>
    /// <remarks>
    /// Returns a list of common initialization file names for C# projects, such as .csproj or
    /// GlobalUsings.cs.
    /// </remarks>
    /// <response code="200">A list of initialization file names.</response>
    [HttpGet]
    [Route("init-files")]
    [ValidateModelState]
    [SwaggerOperation(nameof(GetInitFiles))]
    [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "A list of initialization file names.")]
    public async Task<IActionResult> GetInitFiles() =>
        new ObjectResult(await mediator.Send(new GetInitFiles.Query(SolutionDirectory)));

    /// <summary>
    /// Get supported file extensions
    /// </summary>
    /// <remarks>Returns a list of file extensions recognized by the C# parser.</remarks>
    /// <response code="200">A list of supported file extensions.</response>
    [HttpGet]
    [Route("supported-extensions")]
    [ValidateModelState]
    [SwaggerOperation(nameof(GetSupportedExtensions))]
    [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "A list of supported file extensions.")]
    public IActionResult GetSupportedExtensions()
    {
        logger.LogInformation("Requested extensions");

        return new ObjectResult(new List<string> { "*.cs" });
    }

    /// <summary>
    /// Parse a C# file
    /// </summary>
    /// <remarks>
    /// Returns a normalized ParseResult structure with detailed entities, relationships, and metadata for a given C#
    /// file.
    /// </remarks>
    /// <param name="body"></param>
    /// <response code="200">Successful parsing result.</response>
    /// <response code="400">Invalid request (e.g., malformed JSON or invalid file paths).</response>
    /// <response code="500">Internal server error during parsing.</response>
    [HttpPost]
    [Route("parse")]
    [ValidateModelState]
    [SwaggerOperation(nameof(ParseCsharpFile))]
    [SwaggerResponse(statusCode: 200, type: typeof(ParseResultDto), description: "Successful parsing result.")]
    [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponseDto), description: "Invalid request (e.g., malformed JSON or invalid file paths).")]
    [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponseDto), description: "Internal server error during parsing.")]
    public async Task<IActionResult> ParseCsharpFile([FromBody] ParserRequestDto body)
    {
        if (CheckBasePath(body.RepoPath))
        {
            logger.LogError("Incorrect repo path in request: {Request}", body);
            string message = $"repo directory must be {SolutionDirectory} and nothing else";
            return StatusCode(404, mapper.Map(new ErrorResponse(message, message, null)));
        }

        (ParseResult result, ErrorResponse? error) = await mediator.Send(new ParseCsharpFile.Query(mapper.Map(body)));
        if (error != null)
        {
            return StatusCode(400, mapper.Map(error!));
        }

        return StatusCode(200, mapper.Map(result));
    }

    /// <summary>
    /// Resolve an import path to file paths
    /// </summary>
    /// <remarks>Resolves a C# &#x27;using&#x27; statement to one or more potential relative file paths within the repository.</remarks>
    /// <param name="body"></param>
    /// <response code="200">A list of resolved relative file paths.</response>
    /// <response code="400">Invalid import path or file context.</response>
    [HttpPost]
    [Route("resolve-import")]
    [ValidateModelState]
    [SwaggerOperation(nameof(ResolveImportPath))]
    [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "A list of resolved relative file paths.")]
    [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponseDto), description: "Invalid import path or file context.")]
    public async Task<IActionResult> ResolveImportPath([FromBody] ResolveImportRequestDto body)
    {
        if (CheckBasePath(body.RepoPath))
        {
            logger.LogError("Incorrect repo path in request: {Request}", body);
            string message = $"repo directory must be {SolutionDirectory} and nothing else";
            return StatusCode(404, mapper.Map(new ErrorResponse(message, message, null)));
        }

        (List<string> result, ErrorResponse? error) = await mediator.Send(new ResolveImportPath.Query(mapper.Map(body)));
        if (error != null)
        {
            return StatusCode(400, mapper.Map(error!));
        }

        return StatusCode(200, result);
    }

    private bool CheckBasePath(string repoPath) =>
        String.Compare(
            Path.GetFullPath(SolutionDirectory).TrimEnd('\\').TrimEnd('/'),
            Path.GetFullPath(repoPath).TrimEnd('\\').TrimEnd('/'),
            StringComparison.InvariantCultureIgnoreCase) != 0;
}