# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VelikiyPrikalel.OLLAMACHAT.Web",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/parser/csharp/init-files": {
      "get": {
        "tags": [
          "CParserApi"
        ],
        "summary": "Get initialization file names",
        "description": "Returns a list of common initialization file names for C# projects, such as .csproj or\r\nGlobalUsings.cs.",
        "responses": {
          "200": {
            "description": "A list of initialization file names.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parser/csharp/supported-extensions": {
      "get": {
        "tags": [
          "CParserApi"
        ],
        "summary": "Get supported file extensions",
        "description": "Returns a list of file extensions recognized by the C# parser.",
        "responses": {
          "200": {
            "description": "A list of supported file extensions.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parser/csharp/parse": {
      "post": {
        "tags": [
          "CParserApi"
        ],
        "summary": "Parse a C# file",
        "description": "Returns a normalized ParseResult structure with detailed entities, relationships, and metadata for a given C#\r\nfile.",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful parsing result.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g., malformed JSON or invalid file paths).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during parsing.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parser/csharp/resolve-import": {
      "post": {
        "tags": [
          "CParserApi"
        ],
        "summary": "Resolve an import path to file paths",
        "description": "Resolves a C# 'using' statement to one or more potential relative file paths within the repository.",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ResolveImportRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of resolved relative file paths.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid import path or file context.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OLLAMACHAT.Generated.Models.ErrorResponseDto": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or Sets Code"
          },
          "details": {
            "type": "string",
            "description": "Gets or Sets Details"
          },
          "message": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or Sets Message"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseErrorDto": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.LocationDto"
          },
          "message": {
            "type": "string",
            "description": "Gets or Sets Message"
          },
          "severity": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseErrorDto.SeverityEnumDto"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseErrorDto.SeverityEnumDto": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "Gets or Sets Severity",
        "format": "int32"
      },
      "OLLAMACHAT.Generated.Models.ParseRequest.ParserOptionsDto": {
        "type": "object",
        "properties": {
          "extractFullInheritance": {
            "type": "boolean",
            "description": "Gets or Sets ExtractFullExtractInheritance",
            "nullable": true
          },
          "extractUsingStatementData": {
            "type": "boolean",
            "description": "Gets or Sets ExtractUsingStatementData",
            "nullable": true
          },
          "maxDepth": {
            "type": "integer",
            "description": "Gets or Sets MaxDepth",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Configuration options for selective parsing."
      },
      "OLLAMACHAT.Generated.Models.ParseRequest.ParserRequestDto": {
        "required": [
          "filePath",
          "repoPath"
        ],
        "type": "object",
        "properties": {
          "filePath": {
            "minLength": 1,
            "type": "string",
            "description": "Path to the C# source file to parse, relative to the repository root."
          },
          "options": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseRequest.ParserOptionsDto"
          },
          "repoPath": {
            "minLength": 1,
            "type": "string",
            "description": "Absolute path to the repository root for resolving context."
          }
        },
        "additionalProperties": false,
        "description": "Request to parse a C# file."
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.AttributeDto": {
        "type": "object",
        "properties": {
          "arguments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or Sets Arguments"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.FileMetadataDto": {
        "type": "object",
        "properties": {
          "complexityScore": {
            "type": "integer",
            "description": "Aggregated cyclomatic complexity.",
            "format": "int32",
            "nullable": true
          },
          "loc": {
            "type": "integer",
            "description": "Lines of Code.",
            "format": "int32",
            "nullable": true
          },
          "namespace": {
            "type": "string",
            "description": "The primary namespace of the file."
          }
        },
        "additionalProperties": false,
        "description": "High-level metadata about the parsed file."
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.LocationDto": {
        "required": [
          "end",
          "start"
        ],
        "type": "object",
        "properties": {
          "start": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.PositionDto"
          },
          "end": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.PositionDto"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ModelParameterDto": {
        "type": "object",
        "properties": {
          "defaultValue": {
            "type": "string",
            "description": "Gets or Sets DefaultValue"
          },
          "name": {
            "type": "string",
            "description": "Gets or Sets Name"
          },
          "nullable": {
            "type": "boolean",
            "description": "Gets or Sets Nullable",
            "nullable": true
          },
          "fullTypeName": {
            "type": "string",
            "description": "Gets or Sets FullTypeName"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto": {
        "required": [
          "contentHash",
          "entities",
          "filePath",
          "language",
          "parseTimeMs"
        ],
        "type": "object",
        "properties": {
          "contentHash": {
            "minLength": 1,
            "type": "string",
            "description": "SHA-256 hash of the file content for caching."
          },
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityDto"
            },
            "description": "Gets or Sets Entities"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseErrorDto"
            },
            "description": "Gets or Sets Errors"
          },
          "filePath": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or Sets FilePath"
          },
          "language": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto.LanguageEnumDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.FileMetadataDto"
          },
          "parseTimeMs": {
            "type": "integer",
            "description": "Time taken for parsing in milliseconds.",
            "format": "int32"
          },
          "relationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.RelationshipDto"
            },
            "description": "Gets or Sets Relationships"
          }
        },
        "additionalProperties": false,
        "description": "Complete parsing result for a single file."
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ParseResultDto.LanguageEnumDto": {
        "enum": [
          0
        ],
        "type": "integer",
        "description": "Gets or Sets Language",
        "format": "int32"
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityDto": {
        "required": [
          "location",
          "simpleName",
          "type"
        ],
        "type": "object",
        "properties": {
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityDto"
            },
            "description": "Gets or Sets Children"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.AttributeDto"
            },
            "description": "Gets or Sets Attributes"
          },
          "usingStatementData": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.UsingStatementDataDto"
          },
          "inheritance": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityInheritanceDto"
          },
          "location": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.LocationDto"
          },
          "modifiers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Модификаторы - private/static etc."
          },
          "simpleName": {
            "minLength": 1,
            "type": "string",
            "description": "Gets or Sets SimpleName"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ModelParameterDto"
            },
            "description": "Gets or Sets Parameters"
          },
          "returnType": {
            "type": "string",
            "description": "Gets or Sets ReturnType"
          },
          "fullName": {
            "type": "string",
            "description": "Gets or Sets FullName"
          },
          "type": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityDto.TypeEnumDto"
          }
        },
        "additionalProperties": false,
        "description": "Represents a single parsed code entity (class, method, etc.)."
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityDto.TypeEnumDto": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "description": "Gets or Sets Type",
        "format": "int32"
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.ParsedEntityInheritanceDto": {
        "type": "object",
        "properties": {
          "directBaseClasses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or Sets DirectBaseClasses"
          },
          "directInterfaces": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or Sets DirectInterfaces"
          },
          "allBaseClasses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or Sets AllBaseClasses",
            "nullable": true
          },
          "allInterfaces": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or Sets AllInterfaces",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.PositionDto": {
        "type": "object",
        "properties": {
          "column": {
            "type": "integer",
            "description": "0-based column position.",
            "format": "int32",
            "nullable": true
          },
          "index": {
            "type": "integer",
            "description": "0-based character index from the start of the file.",
            "format": "int32",
            "nullable": true
          },
          "line": {
            "type": "integer",
            "description": "1-based line number.",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.RelationshipDto": {
        "required": [
          "fullNameFrom",
          "fullNameTo",
          "type"
        ],
        "type": "object",
        "properties": {
          "fullNameFrom": {
            "minLength": 1,
            "type": "string",
            "description": "Full name of the source entity."
          },
          "targetDefinitionFilePath": {
            "type": "string",
            "description": "Relative path to the target file if the relationship is cross-file."
          },
          "fullNameTo": {
            "minLength": 1,
            "type": "string",
            "description": "Full name of the target entity."
          },
          "type": {
            "$ref": "#/components/schemas/OLLAMACHAT.Generated.Models.ParseResponse.RelationshipDto.TypeEnumDto"
          }
        },
        "additionalProperties": false,
        "description": "Represents a relationship between two entities."
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.RelationshipDto.TypeEnumDto": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "Gets or Sets Type",
        "format": "int32"
      },
      "OLLAMACHAT.Generated.Models.ParseResponse.UsingStatementDataDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "description": "Gets or Sets Source"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "OLLAMACHAT.Generated.Models.ResolveImportRequestDto": {
        "required": [
          "filePath",
          "importPath",
          "repoPath"
        ],
        "type": "object",
        "properties": {
          "filePath": {
            "minLength": 1,
            "type": "string",
            "description": "The path of the file containing the import."
          },
          "importPath": {
            "minLength": 1,
            "type": "string",
            "description": "The import/using path to resolve."
          },
          "repoPath": {
            "minLength": 1,
            "type": "string",
            "description": "Absolute path to the repository root."
          }
        },
        "additionalProperties": false,
        "description": ""
      }
    }
  }
}